refactor: move functions to new file (no changes
